// WorkSpace/projects/user/src/app/models/Service/iservice-read.ts
export interface IServiceRead {
  id: number;
  name: string;
  description: string;
  image: string;
  dateTime: Date;
  companyId: number;
  companyName: string;
  companySlogan: string;
  companyEmployees: number;
  companyImage: string;
  companyEmail: string;
  companyPhone: string;
  companyLink: string;
}